backend: create unified BTC/LTC accounts, with option for splitting
`createAndAddBTCAccount()` is added and used to create Bitcoin and
Litecoin accounts, combining the Legacy, Segwit and Bech32
subaccounts.

The keystore is checked for support of each subtype and the account
skipped if none are supported. For example, LTC is skipped completely
on the BitBox02 Bitcoin-only, and Legacy is skipped for all BitBox02s.

The old way of showing one account per address/script type is
preserved for the BitBox01 (does not support mixed tx inputs) or if
the user opts-in via the expert setting.