backend: remove multisig
It was a "hacky" 2-of-2 multisig if two bitboxes are connected, which
was used to play with multisig in the past. The multisig argument to
enable this mode has been long removed. The cosigner index concept was
also wrong, as the cosigner index is not a property of the keystore,
but a property of the signing configuration (and could be infered
there from the xpubs that are present).