backend: modify accounts config atomically
Before, the ModifyAccountsConfig() function would be called many
times, e.g. for each account, for what is actually one operation.

This changes the use to be more like a transactional DB, where you
only pass a reference to the DB around, and do the actual
openining/committing at a higher level.

This is needed so that we can atomically determine the next account
number and also add the new account, reusing the existing utility
functions such as persistBTCAccountConfig and filterAccounts.

The config locks have been split into one for the app config and one
for the account config, as they are independent: you want to be able
to read the app config while modifying the accounts config.