qt: enforce a single instance, forward args to the primary instance
This library seems to do everything we want in the same way we would
do it manully, by using shared memory:
https://github.com/itay-grudev/SingleApplication

We want to enforce a single instance because the app behaves badly if
there are multiple instances, e.g. the 2nd one blocks on files opened
by the first instance, etc.

We also want to handle URI protocols while the app is already
running. For Linux and Windows, a new instance is launched with the
URI as the CLI argument. We forward this to the primary instance to
handle.