backend/accounts: add types package to break cyclic deps
The plan is to remove lots of duplicate fields of
accounts.AccountsConfig and config.Account by using config.Account in
accounts.AccountsConfig, but doing that leads to cyclic deps because
both depend on accounts.Event and accounts.Code. This makes a new
types package to break the cycle, starting with Event.