frontend/account: simplify coin unit types
Rename `Coin` to `CoinUnit`, as it is the units, the the coin codes or
similar.

Delete/inline the other definitions as they are only used there.

IAmount:unit should have included the sat units all along as it
produced by `btc/handlers/formatAmountAsJSON`, which calls
`accountCoin.GetFormatUnit(isFee)`, which returns `sat/tsat` for BTC
in sats mode.