frontend/receive: fix receive screen for software keystore
In the receive screen of accounts using the software keystore, the
address and QR code have been show upfront without needing to click a
verify button.

In watch-only, we don't know for sure which kind of keystore is behind
the account until it is connected. So we always let the user click the
verify button and react accordingly if it is a software keystore.

The label on the button says 'Verify on BitBox02' - we leave that even
for the software keystore as changing it requires re-translation of
that i18n key in all languages. The software keystore is not used in
production, so we can live with this for developers for the time
being.

The BitBox01 has its own receive-bb01.tsx component independent of
this change.