ci: refactor macOS build
We want to call `make ios` in another step, but using `github-ci.sh`
again for this like for Qt would repeat all the steps related to
installing Brew, setting up Go, coping the repo to $GOPATH, etc.

This commit refactors it so these steps are done upfront in separate
steps, so we can append a `make ios` step afterwards.