ci: build iOS app
For now we ignore signing the app, publishing it or even running
it. This first step is to simply compile the iOS app to check for
build errors.

In the future, we may want to consider using github secrets to
automatically sign and push to testflight.

The macos runner image is bumped to 14, which defaults t XCode 15. The
older version had some build error.